+++ mobile/android/android-components/components/lib/crash/src/main/java/mozilla/components/lib/crash/service/CrashReporterService.kt
@@ -6,6 +6,8 @@ package mozilla.components.lib.crash.service
 
 import mozilla.components.lib.crash.Crash
 
+internal const val INFO_PREFIX = "[INFO]"
+
 /**
  * Interface to be implemented by external services that accept crash reports.
  */
@@ -19,4 +21,9 @@ interface CrashReporterService {
      * Submits a crash report for this [Crash.NativeCodeCrash].
      */
     fun report(crash: Crash.NativeCodeCrash)
+
+    /**
+     * Submits a caught exception report for this [Throwable].
+     */
+    fun report(throwable: Throwable)
 }